+++ wbooks/tsconfig.json
@@ -67,7 +67,8 @@
       "@services/*":["src/services/*"],
       "@utils":["src/utils"],
       "@navigator":["src/app/navigator"],
-      "@redux/*":["src/redux/*"]
+      "@redux/*":["src/redux/*"],
+      "@reduxtest/*":["../../../src/redux/*"]
     },
   },
   "include": ["index.d.ts", "src/**/*", "src/App.test.tsx"],
